Adding a reference Daemon-CLI application
Basic Functionality:
1) Have a CLI from which a user can select from pre_determined tasks
2) a Daemon running in the background, waiting for this input and boiler code
for handling the different tasks in different threads
3) Once the thread task is complete, it should go away. But daemon should keep running

daemon.rs and cli.rs: get the above functionality

Update:
the CLI code to send the request to the daemon and wait for a response from the daemon.
Then print the output that response is received and then the CLI process closes.
daemon_bidirec.rs and cli_bidirec.rs: for the above

Update:
Using tokio crate for async processing
daemon_tokio.rs and cli_tokio.rs: for the above